Move TableItem to Table directory
