+++ src/components/Table/TableItem/styles.module.scss
@@ -1,4 +1,4 @@
-@import "~$styles/variables";
+@import "../../../../node_modules/$styles/variables";
 
 .tableItem {
   background-color: $white;
